Break store into modules
